meme_application/requirements.txt

python-dotenv
google-cloud-storage
google-cloud-bigquery
google-cloud-texttospeech
google-cloud-translate
google-cloud-vision
Flask
Pillow==9.5.0
google-cloud-datastore
google-cloud-aiplatform
langchain
black